requirements.txt
flask>=3.0.3
flask-cors>=5.0.0
flask-pymongo>=2.3.0
vietocr>=0.3.13
ultralytics>=8.2.90
python-dotenv>=1.0.1